Test:  Unit Tests for Uploading Datasets Zenodo
Two unit tests have been created: one to test the function that creates a dataset and uploads it directly to Zenodo, and another to test the function that uploads a dataset to Zenodo from the staging area.

+++ app/modules/dataset/tests/test_unit.py
@@ -79,8 +79,10 @@ def client():
             yield client
 
             # Limpiar el archivo temporal después de la prueba
-            os.remove(os.path.join(temp_folder, 'file9.uvl'))
-            os.rmdir(temp_folder)
+            if os.path.exists(os.path.join(temp_folder, 'file9.uvl')):
+                os.remove(os.path.join(temp_folder, 'file9.uvl'))
+            if os.path.exists(temp_folder):
+                os.rmdir(temp_folder)
 
             db.session.remove()
             db.drop_all()
@@ -123,6 +125,51 @@ def test_create_dataset(client):
 
     logout(client)
     
+def test_upload_dataset_zenodo(client):
+    login_response = login(client, "user5@example.com", "1234")
+    assert login_response.status_code == 200, "Login was unsuccessful."
+
+    # Datos de ejemplo para crear el dataset
+    form_data = {
+        "title": "test dataset in zenodo",
+        "desc": "test description",
+        "publication_type": "none",
+        "publication_doi": "",
+        "dataset_doi": "",
+        "tags": "",
+        "authors-0-name": "Author Name",
+        "authors-0-affiliation": "Author Affiliation",
+        "authors-0-orcid": "0000-0001-2345-6789",
+        "feature_models-0-uvl_filename": "file9.uvl",
+        "feature_models-0-title": "Feature Model Title",
+        "feature_models-0-desc": "Feature Model Description",
+        "feature_models-0-publication_type": "none",
+        "feature_models-0-publication_doi": "",
+        "feature_models-0-tags": "",
+        "feature_models-0-version": "1.0",
+        "feature_models-0-authors-0-name": "FM Author Name",
+        "feature_models-0-authors-0-affiliation": "FM Author Affiliation",
+        "feature_models-0-authors-0-orcid": "0000-0002-3456-7890"
+    }
+
+    # Enviar la solicitud POST con los datos del formulario para crear el dataset
+    response = client.post('/dataset/upload', data=form_data)
+    assert response.status_code == (200 or 403), f"Expected status code 200 or 403, but got {response.status_code}"
+
+    # Verificar que el dataset se haya creado correctamente
+    dataset = DataSet.query.join(DSMetaData).filter(DSMetaData.title == "test dataset in zenodo").first()
+    assert dataset is not None, "Dataset was not created"
+
+    # Hacer una solicitud GET a la ruta /dataset/list para verificar que el dataset se encuentra en el listado de datasets
+    response = client.get('/dataset/list')
+    assert response.status_code == 200, f"Expected status code 200, but got {response.status_code}"
+
+    # Verificar que el dataset se encuentra en el listado de datasets
+    html_data = response.data.decode('utf-8')
+    assert "test dataset in zenodo" in html_data, "The created dataset is not in the list of datasets"
+
+    logout(client)
+
 def test_create_and_list_unprepared_dataset(client):
     login_response = login(client, "user5@example.com", "1234")
     assert login_response.status_code == 200, "Login was unsuccessful."
@@ -160,7 +207,7 @@ def test_create_and_list_unprepared_dataset(client):
     response = client.get('/dataset/list')
     assert response.status_code == 200, f"Expected status code 200, but got {response.status_code}"
 
-    # Verificar que el dataset creado esté en los unprepared_datasets
+    # Verificar que el dataset esta en el listado de datasets
     html_data = response.data.decode('utf-8')
     assert "test" in html_data, "The created dataset is not in the unprepared_datasets"
 
@@ -212,6 +259,57 @@ def test_update_staging_area_dataset(client):
     assert updated_dataset.ds_meta_data.description == "updated description", f"Expected description 'updated description', but got {updated_dataset.ds_meta_data.description}"
     assert updated_dataset.ds_meta_data.authors[0].name == "TestSurname, TestName", f"Expected author name 'TestSurname, TestName', but got {updated_dataset.ds_meta_data.authors[0].name}"
     assert updated_dataset.ds_meta_data.authors[1].name == "Updated Author Name", f"Expected author name 'Updated Author Name', but got {updated_dataset.ds_meta_data.authors[0].name}"
+
+    logout(client)
+    
+def test_upload_dataset_zenodo_from_staging(client):
+    login_response = login(client, "user5@example.com", "1234")
+    assert login_response.status_code == 200, "Login was unsuccessful."
+
+    dataset = DataSet.query.join(DSMetaData).filter(DSMetaData.title == "Staging area Dataset").first()
+    assert dataset is not None, "Dataset with title 'Staging area Dataset' not found"
+    dataset_id = dataset.id
+
+    # Datos de ejemplo para actualizar el dataset
+    form_data_update = {
+        "title": "updated unique dataset",
+        "desc": "updated description",
+        "publication_type": "none",
+        "publication_doi": "",
+        "dataset_doi": "",
+        "tags": "",
+        "authors-0-name": "Updated Author Name",
+        "authors-0-affiliation": "Updated Author Affiliation",
+        "authors-0-orcid": "0000-0001-2345-6789",
+        "feature_models-0-uvl_filename": "file9.uvl",
+        "feature_models-0-title": "Updated Feature Model Title",
+        "feature_models-0-desc": "Updated Feature Model Description",
+        "feature_models-0-publication_type": "none",
+        "feature_models-0-publication_doi": "",
+        "feature_models-0-tags": "",
+        "feature_models-0-version": "1.0",
+        "feature_models-0-authors-0-name": "Updated FM Author Name",
+        "feature_models-0-authors-0-affiliation": "Updated FM Author Affiliation",
+        "feature_models-0-authors-0-orcid": "0000-0002-3456-7890",
+        "csrf_token": "ImM4Y2YwNzA1NGYxNGZhMDlhMjE3ZjQ4NDRjMTVkNTc5ZmY5ZWQwMjQi.Z0tv3Q.IA3Me8feSsE3T5juVZT6k9StAj4"
+    }
+
+    # Enviar la solicitud POST con los datos del formulario para actualizar el dataset
+    response = client.post(f'/dataset/upload/{dataset_id}', data=form_data_update)
+    
+    # Espera 200 en caso de que se suba satisfactoramente a zenodo o 403 en caso de que zenodo rechace la conexion
+    assert response.status_code == (200 or 403), f"Expected status code 200 or 403, but got {response.status_code}"
+    
+    # Se espera que tenga el valor de False en staging_area
+    assert dataset.ds_meta_data.staging_area == False, f"Expected staging_area False, but got {dataset.staging_area}"
+
+    # Hacer una solicitud GET a la ruta /dataset/list para verificar que el dataset se encuentra en la sección local_datasets
+    response = client.get('/dataset/list')
+    assert response.status_code == 200, f"Expected status code 200, but got {response.status_code}"
+
+    # Verificar que el dataset esta en el listado de datasets
+    html_data = response.data.decode('utf-8')
+    assert "updated unique dataset" in html_data, "The updated dataset is not in the local_datasets"
 
     logout(client)
